requirements.txt
Python 3.7 or greater
The following libraries:
Pillow
pytesseract
tesseract-ocr
ephem
urllib.request
re
math
sys